+++ gwcelery/tests/test_tasks_p_astro_gstlal.py
@@ -1,6 +1,5 @@
 import json
 from unittest.mock import patch
-from urllib import request
 
 import numpy as np
 import pkg_resources
@@ -10,17 +9,29 @@ from .. import app
 from ..tasks import p_astro_gstlal
 
 
-class MockResponse(object):
+class MockResponseWeights(object):
+    def __init__(self):
+        self.url = "data/H1L1V1-weights-bins_686-1126051217-61603201.json"
+
     def read(self):
-        return json.dumps({
-            "counts_BNS": 2,
-            "counts_BBH": 10,
-            "counts_NSBH": 1,
-            "counts_Terrestrial": 4000,
-            "counts_MassGap": 1})
+        with pkg_resources.resource_stream(__name__, self.url) as f:
+            return f.read()
 
     def close(self):
         pass
+
+
+@pytest.fixture
+def mock_url(monkeypatch):
+    def _urlfunc(url):
+        filename = "data/H1L1V1-weights-bins_686-1126051217-61603201.json"
+        if url == app.conf['p_astro_url']:
+            with pkg_resources.resource_stream(__name__, filename) as f:
+                return f
+        elif url == app.conf['p_astro_weights_url']:
+            return MockResponseWeights()
+
+    monkeypatch.setattr('urllib.request.urlopen', _urlfunc)
 
 
 @pytest.fixture
@@ -65,23 +76,23 @@ def test_get_ln_f_over_b(ranking_data_bytes):
 
 
 def test_get_event_ln_likelihood_ratio_svd_endtime_mass(coinc_bytes_1):
-    likelihood, end_time, mass, mass1, mass2, snr, far = \
+    likelihood, mass1, mass2, spin1z, spin2z, snr, far = \
         p_astro_gstlal._get_event_ln_likelihood_ratio_svd_endtime_mass(
             coinc_bytes_1)
     assert mass1 == pytest.approx(2.8, abs=0.1)
     assert mass2 == pytest.approx(1.0, abs=0.1)
+    assert spin1z == pytest.approx(-0.99, abs=0.01)
+    assert spin2z == pytest.approx(0.049, abs=0.01)
     assert likelihood == pytest.approx(21.65, abs=0.1)
-    assert end_time == pytest.approx(1174052512)
-    assert mass == pytest.approx(3.78, abs=0.1)
 
 
-@patch.object(request, 'urlopen', return_value=MockResponse())
-def test_compute_p_astro_1(mock_url, coinc_bytes_1, ranking_data_bytes):
+def test_compute_p_astro_1(coinc_bytes_1, ranking_data_bytes, mock_url):
     """Test to call `compute_p_astro` on gracedb event G322589.
     m1 = 2.7, m2 = 1.0 solar mass for this event"""
     files = coinc_bytes_1, ranking_data_bytes
-    # FIXME graceid should be removed once fixed
+
     p_astros = json.loads(p_astro_gstlal.compute_p_astro(files))
+
     assert p_astros['BNS'] == pytest.approx(1, abs=1e-2)
     assert p_astros['NSBH'] == pytest.approx(0, abs=1e-2)
     assert p_astros['BBH'] == pytest.approx(0, abs=1e-2)
@@ -89,19 +100,18 @@ def test_compute_p_astro_1(mock_url, coinc_bytes_1, ranking_data_bytes):
     assert p_astros['Terrestrial'] == pytest.approx(0, abs=1e-2)
 
 
-@patch.object(request, 'urlopen', return_value=MockResponse())
-def test_compute_p_astro_2(mock_url, coinc_bytes_2, ranking_data_bytes):
+def test_compute_p_astro_2(coinc_bytes_2, ranking_data_bytes, mock_url):
     """Test to call `compute_p_astro` on gracedb event G5351.
-    m1 = 4.2, m2 = 1.2 solar mass for this event. FAR = 1.9e-6, P_terr
+    m1 = 1.1, m2 = 1.0 solar mass for this event. FAR = 1.9e-6, P_terr
     has a moderate value."""
     files = coinc_bytes_2, ranking_data_bytes
-    # FIXME graceid should be removed once fixed
+
     p_astros = json.loads(p_astro_gstlal.compute_p_astro(files))
+
     assert p_astros['Terrestrial'] > 0.30
 
 
-@patch.object(request, 'urlopen', return_value=MockResponse())
-def test_failing_compute_p_astro(mock_url, coinc_bytes_3, ranking_data_bytes):
+def test_failing_compute_p_astro(coinc_bytes_3, ranking_data_bytes, mock_url):
     """Test the case when p_astro computation fails"""
     files = coinc_bytes_3, ranking_data_bytes
     with patch('gwcelery.tasks.p_astro_other.compute_p_astro') as p: